Changed database route

+++ app/models/Room.scala
@@ -7,17 +7,17 @@ import org.mongodb.scala.bson.ObjectId
 import org.mongodb.scala.bson.codecs.DEFAULT_CODEC_REGISTRY
 import play.api.libs.json.{JsString, JsValue, Json, Writes}
 
-case class Room(_id:ObjectId, room_id: Double, hotel_id: Double, room_type:String,
-                city:String, capacity: Double, price:Double, currency:String,
+case class Room(_id:ObjectId, room_id: Int, hotel_id: Int, room_type:String,
+                city:String, capacity: Int, price:Int, currency:String,
                 description:String)
 
 object Room{
 
-  def apply( room_id: Double, hotel_id: Double, room_type:String,
-             city:String,capacity: Double, price:Double, currency:String,
+  def apply( room_id: Int, hotel_id: Int, room_type:String,
+             city:String, capacity: Int, price:Int, currency:String,
              description:String): Room =
-    Room(new ObjectId(),room_id: Double, hotel_id: Double, room_type:String,
-      city:String,capacity: Double, price:Double, currency:String,
+    Room(new ObjectId(),room_id: Int, hotel_id: Int, room_type:String,
+      city:String, capacity: Int, price:Int, currency:String,
       description:String)
 
   //Para poder convertir la colección en BSON y viceversa
